refactor: Move tape_builder.py to onnxscript/_internal/
Consistent with the placement of builder.py and other internal modules.
Public exports from onnxscript.__init__ remain unchanged.
